Support switching between Piwik versions

+++ generate.sh
@@ -2,11 +2,11 @@
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $DIR
-git checkout docs/generated
-rm -rf docs/cache
-mkdir docs/cache
-rm -rf docs/generated
-mkdir docs/generated
+git checkout docs/*/generated
+rm -rf docs/*/cache
+mkdir docs/*/cache
+rm -rf docs/*/generated
+mkdir docs/*/generated
 cd piwik
 git reset --hard
 git submodule foreach --recursive git reset --hard
@@ -18,7 +18,15 @@ git pull origin master
 git submodule update --recursive --force
 php composer.phar install || true
 cd ..
-php generator/generate.php
+php generator/generate.php --branch=master --targetname=2.x
+GENERATION_SUCCESS=$?
+
+if [ $GENERATION_SUCCESS -ne 0 ]; then
+  exit 1;
+fi
+
+php generator/generate.php --branch=3.x-dev --targetname=3.x
+
 GENERATION_SUCCESS=$?
 rm -rf app/tmp/cache/*
